Organized team members file into api route
-Same as chapters, created a new api route for team members, updated and moved necessary files into it
