--- src/app/about/Init_Member.interface.ts
@@ -1,6 +1,0 @@
-export interface Init_Member {
-  name: string;
-  position: string;
-  linkedin: string;
-  image: string;
-}
