+++ src/app/api/eboard/InitTeam.ts
@@ -1,51 +1,46 @@
-export const team_members = [
+import { Member } from "./route";
+
+export const team_members: Member[] = [
   {
     name: "Julieta Vallejos",
     position: "Executive Director, INIT",
     linkedin: "https://www.linkedin.com/in/julieta-vallejos-pasco-61985790/",
-    image:
-      "/julieta_vallejos.jpg",
+    image: "/images/eboard/julieta_vallejos.jpg",
   },
   {
     name: "Cesar Villa-Garcia",
     position: " Product Manager, Microsoft",
     linkedin: "https://www.linkedin.com/in/cesar-villa-garcia/",
-    image:
-      "/cesar_villa-garcia.jpeg",
+    image: "/images/eboard/cesar_villa-garcia.jpeg",
   },
   {
     name: "Yasmine Abdrabo",
     position: " Site Reliability Engineer, Citi",
     linkedin: "https://www.linkedin.com/in/yasmine-abdrabo/",
-    image:
-      "/yasmine_abdrabo.jpeg",
+    image: "/images/eboard/yasmine_abdrabo.jpeg",
   },
   {
     name: "Jose Maldonado",
     position: " Software Engineer, Xbox",
     linkedin: "https://www.linkedin.com/in/jmald045/",
-    image:
-      "/jose_maldonado.jpeg",
+    image: "/images/eboard/jose_maldonado.jpeg",
   },
   {
     name: "Sabina Cartacio",
     position: " Product Manager, Fiddler",
     linkedin: "https://www.linkedin.com/in/cartacios/",
-    image:
-      "/sabina_cartacio.jpeg",
+    image: "/images/eboard/sabina_cartacio.jpeg",
   },
   {
     name: "Christopher Rodriguez",
     position: " Software Engineer, Microsoft",
     linkedin: "https://www.linkedin.com/in/chrisrxyz/",
-    image:
-      "/christopher_rodriguez.jpeg",
+    image: "/images/eboard/christopher_rodriguez.jpeg",
   },
   {
     name: "Andy Garcia",
     position: " INIT FIU Chapter President",
     linkedin: "https://www.linkedin.com/in/andygarcia-tech/",
-    image:
-      "/andy_garcia.jpeg",
+    image: "/images/eboard/andy_garcia.jpeg",
   },
 ];
